requirements.txt
arrow==1.2.3
asgiref==3.7.2
binaryornot==0.4.4
certifi==2023.5.7
chardet==5.1.0
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
contourpy==1.0.7
cookiecutter==2.1.1
cycler==0.11.0
Django==4.2.1
django-browser-reload==1.8.0
django-tailwind==3.5.0
fontawesomefree==6.4.0
fonttools==4.39.4
idna==3.4
Jinja2==3.1.2
jinja2-time==0.2.0
joblib==1.2.0
kiwisolver==1.4.4
MarkupSafe==2.1.3
matplotlib==3.7.1
mysql-connector-python==8.0.33
mysqlclient==2.1.1
numpy==1.24.3
opencv-python==4.7.0.72
packaging==23.1
Pillow==9.5.0
protobuf==3.20.3
pyparsing==3.0.9
python-dateutil==2.8.2
python-slugify==8.0.1
PyYAML==6.0
requests==2.31.0
scikit-learn==1.2.2
scipy==1.10.1
six==1.16.0
sqlparse==0.4.4
sweetify==2.3.1
text-unidecode==1.3
threadpoolctl==3.1.0
tzdata==2023.3
urllib3==2.0.2